fix : git ignorecase false
